perf: improves StencilLayer depth calculation

+++ source/Age/Passes/Scene2DPass.cs
@@ -167,7 +167,7 @@ public abstract class Scene2DPass : RenderPass<Texture2D>
             {
                 if (command.StencilLayer is StencilLayer currentLayer)
                 {
-                    var currentDepth = currentLayer.Depth + 1;
+                    var currentDepth = (uint)(currentLayer.Depth + 1);
 
                     while (this.stencilStack.Count > currentDepth)
                     {
